+++ block/internal/cache/pending_headers.go
@@ -80,10 +80,6 @@ func (ph *PendingHeaders) SetLastSubmittedHeaderHeight(ctx context.Context, newL
 	ph.base.setLastSubmittedHeight(ctx, newLastSubmittedHeaderHeight)
 }
 
-func (ph *PendingHeaders) init() error {
-	return ph.base.init()
-}
-
 func (ph *PendingHeaders) GetLastSubmittedHeaderHeight() uint64 {
 	return ph.base.getLastSubmittedHeight()
 }